+++ mappraiser/src/domain_generalization.c
@@ -33,7 +33,7 @@ int initialize_PCG_var_struct(PCG_var *PCG_variable, double *local_map_pix)
 }
 
 
-int init_harmonic_superstruct(Mat *A, Harmonic_superstruct *Harm_struct, int *mask_binary, int nside, int lmax, char *c_ell_path, int number_correlations)
+int init_harmonic_superstruct(Mat *A, Harmonic_superstruct *Harm_struct, double *mask_binary, int nside, int lmax, char *c_ell_path, int number_correlations)
 {
     /* Initalize all structures necessary for harmonic structures : S2HAT_params for S2HAT operations, and the 
        Relies on the fact that nstokes == A->nnz
